+++ internal/cmd/apply.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"fmt"
 	"os/exec"
+	"runtime"
 	"strings"
 
 	"github.com/spf13/cobra"
@@ -13,7 +14,6 @@ import (
 	"github.com/craftcms/nitro/internal/nitro"
 	"github.com/craftcms/nitro/internal/runas"
 	"github.com/craftcms/nitro/internal/scripts"
-
 	"github.com/craftcms/nitro/internal/task"
 	"github.com/craftcms/nitro/internal/webroot"
 )
@@ -123,6 +123,11 @@ var applyCommand = &cobra.Command{
 		if flagSkipHosts || len(configFile.Sites) == 0 {
 			fmt.Println("Skipping editing the hosts file.")
 			return nil
+		}
+
+		// if this is windows, its always root - we should clean this up eventually
+		if runtime.GOOS == "windows" {
+			return hostsCommand.RunE(cmd, args)
 		}
 
 		return runas.Elevated(machine, []string{"hosts"})